refactor: combine all the flags related code into a file
