+++ src/commands/monika.ts
@@ -22,7 +22,7 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { Command, Errors, Flags } from '@oclif/core'
+import { Command, Errors } from '@oclif/core'
 import pEvent from 'p-event'
 
 import type { Config } from '../interfaces/config'
@@ -32,38 +32,27 @@ import { createConfig, getConfig, isSymonModeFrom } from '../components/config'
 import { sortProbes } from '../components/config/sort'
 import { printAllLogs } from '../components/logger'
 import { flush } from '../components/logger/flush'
-import { closeLog, openLogfile } from '../components/logger/history'
+import { closeLog } from '../components/logger/history'
 import { logStartupMessage } from '../components/logger/startup-message'
 import { scheduleSummaryNotification } from '../components/notification/schedule-notification'
 import { sendMonikaStartMessage } from '../components/notification/start-message'
+import { printSummary } from '../components/summary'
 import { getContext, setContext } from '../context'
 import events from '../events'
-import {
-  type MonikaFlags,
-  monikaFlagsDefaultValue,
-  retryInitialDelayMs,
-  retryMaxDelayMs,
-  symonAPIVersion,
-  symonGetProbesIntervalMs,
-} from '../flag'
-import { printSummary, savePidFile } from '../jobs/summary-notification'
+import { type MonikaFlags, sanitizeFlags, flags } from '../flag'
+import { savePidFile } from '../jobs/summary-notification'
 import initLoaders from '../loaders'
 import { sanitizeProbe, startProbing } from '../looper'
 import SymonClient from '../symon'
 import { getEventEmitter } from '../utils/events'
 import { log } from '../utils/pino'
-
-type GetProbesParams = {
-  config: Config
-  flags: MonikaFlags
-}
+import { fetchAndCacheNetworkInfo } from '../utils/public-ip'
 
 const em = getEventEmitter()
 let symonClient: SymonClient
 
 export default class Monika extends Command {
   static description = 'Monika command line monitoring tool'
-
   static examples = [
     'monika',
     'monika --logs',
@@ -72,208 +61,7 @@ export default class Monika extends Command {
     'monika --config https://raw.githubusercontent.com/hyperjumptech/monika/main/monika.example.yml --config-interval 900',
   ]
 
-  static flags = {
-    'auto-update': Flags.string({
-      description:
-        'Enable auto-update for Monika. Available options: major, minor, patch. This will make Monika terminate itself on successful update but does not restart',
-    }),
-    config: Flags.string({
-      char: 'c',
-      default: monikaFlagsDefaultValue.config,
-      description:
-        'JSON configuration filename or URL. If none is supplied, will look for monika.yml in the current directory',
-      env: 'MONIKA_JSON_CONFIG',
-      multiple: true,
-    }),
-
-    'config-filename': Flags.string({
-      default: monikaFlagsDefaultValue['config-filename'],
-      dependsOn: ['config'],
-      description:
-        'The configuration filename for config file created if there is no config file found ',
-    }),
-
-    'config-interval': Flags.integer({
-      default: monikaFlagsDefaultValue['config-interval'],
-      dependsOn: ['config'],
-      description:
-        'The interval (in seconds) for periodic config checking if url is used as config source',
-    }),
-
-    'create-config': Flags.boolean({
-      description:
-        'Create config from HAR (-H), postman (-p), insomnia (-I), sitemap (--sitemap), textfile (--text) export file, or open Monika Configuration Generator using default browser',
-    }),
-
-    'native-fetch': Flags.boolean({
-      default: monikaFlagsDefaultValue['native-fetch'],
-      description:
-        'Use native fetch Node.js API instead of Axios for HTTP client',
-    }),
-
-    flush: Flags.boolean({
-      description: 'Flush logs',
-    }),
-
-    'follow-redirects': Flags.integer({
-      default: monikaFlagsDefaultValue['follow-redirects'],
-      description:
-        'Monika will follow redirects as many times as the specified value here. By default, Monika will follow redirects once. To disable redirects following, set the value to zero.',
-    }),
-
-    force: Flags.boolean({
-      default: false,
-      description: 'Force commands with a yes whenever Y/n is prompted.',
-    }),
-
-    har: Flags.string({
-      char: 'H', // (H)ar file to
-      description: 'Run Monika using a HAR file',
-      exclusive: ['postman', 'insomnia', 'sitemap', 'text'],
-      multiple: false,
-    }),
-
-    help: Flags.help({ char: 'h' }),
-
-    id: Flags.string({
-      char: 'i', // (i)ds to run
-      description: 'Define specific probe ids to run',
-      multiple: false,
-    }),
-
-    ignoreInvalidTLS: Flags.boolean({
-      description:
-        'Configures whether HTTPS requests should ignore invalid certificates',
-      default: false,
-    }),
-
-    insomnia: Flags.string({
-      char: 'I', // (I)nsomnia file to
-      description: 'Run Monika using an Insomnia json/yaml file',
-      exclusive: ['har', 'postman', 'sitemap', 'text'],
-      multiple: false,
-    }),
-
-    'keep-verbose-logs': Flags.boolean({
-      default: false,
-      description: 'Store all requests logs to database',
-    }),
-
-    logs: Flags.boolean({
-      char: 'l', // prints the (l)ogs
-      description: 'Print all logs.',
-    }),
-
-    'one-probe': Flags.boolean({
-      dependsOn: ['sitemap'],
-      description: 'One Probe',
-    }),
-
-    output: Flags.string({
-      char: 'o', // (o)utput file to write config to
-      description: 'Write monika config file to this file',
-      multiple: false,
-    }),
-
-    postman: Flags.string({
-      char: 'p', // (p)ostman
-      description: 'Run Monika using a Postman json file.',
-      exclusive: ['har', 'insomnia', 'sitemap', 'text'],
-      multiple: false,
-    }),
-
-    prometheus: Flags.integer({
-      description:
-        'Specifies the port the Prometheus metric server is listening on. e.g., 3001. (EXPERIMENTAL)',
-      exclusive: ['r'],
-    }),
-
-    repeat: Flags.integer({
-      char: 'r', // (r)epeat
-      default: 0,
-      description: 'Repeats the test run n times',
-      multiple: false,
-    }),
-
-    retryInitialDelayMs,
-
-    retryMaxDelayMs,
-
-    sitemap: Flags.string({
-      description: 'Run Monika using a Sitemap xml file.',
-      exclusive: ['har', 'insomnia', 'postman', 'text'],
-      multiple: false,
-    }),
-
-    'status-notification': Flags.string({
-      description: 'Cron syntax for status notification schedule',
-    }),
-
-    stun: Flags.integer({
-      char: 's', // (s)stun
-      default: monikaFlagsDefaultValue.stun,
-      description: 'Interval in seconds to check STUN server',
-      multiple: false,
-    }),
-
-    summary: Flags.boolean({
-      default: false,
-      description: 'Display a summary of monika running stats',
-    }),
-
-    'symon-api-version': symonAPIVersion(),
-
-    symonKey: Flags.string({
-      dependsOn: ['symonUrl'],
-      description: 'API Key for Symon',
-    }),
-
-    symonGetProbesIntervalMs,
-
-    symonLocationId: Flags.string({
-      dependsOn: ['symonKey', 'symonUrl'],
-      description: 'Location ID for Symon (optional)',
-      required: false,
-    }),
-
-    symonMonikaId: Flags.string({
-      dependsOn: ['symonKey', 'symonUrl'],
-      description: 'Monika ID for Symon (optional)',
-      required: false,
-    }),
-
-    symonReportInterval: Flags.integer({
-      dependsOn: ['symonKey', 'symonUrl'],
-      description: 'Interval for reporting to Symon in milliseconds (optional)',
-      required: false,
-    }),
-
-    symonReportLimit: Flags.integer({
-      dependsOn: ['symonKey', 'symonUrl'],
-      description: 'Data limit to be reported to Symon (optional)',
-      required: false,
-    }),
-
-    symonUrl: Flags.string({
-      dependsOn: ['symonKey'],
-      description: 'URL of Symon',
-      hidden: false,
-    }),
-
-    text: Flags.string({
-      description: 'Run Monika using a Simple text file',
-      exclusive: ['postman', 'insomnia', 'sitemap', 'har'],
-      multiple: false,
-    }),
-
-    verbose: Flags.boolean({
-      default: false,
-      description: 'Show verbose log messages',
-    }),
-
-    version: Flags.version({ char: 'v' }),
-  }
-
+  static flags = flags
   static id = 'monika'
 
   async catch(error: Error): Promise<unknown> {
@@ -297,113 +85,63 @@ export default class Monika extends Command {
     throw error
   }
 
-  deprecationHandler(config: Config): Config {
-    const showDeprecateMsg: Record<'query', boolean> = {
-      query: false,
-    }
-
-    const checkedConfig = {
-      ...config,
-      probes: config.probes?.map((probe) => ({
-        ...probe,
-        requests: probe.requests?.map((request) => ({
-          ...request,
-          alert: request.alerts?.map((alert) => {
-            if (alert.query) {
-              showDeprecateMsg.query = true
-              return { ...alert, assertion: alert.query }
-            }
-
-            return alert
-          }),
-        })),
-        alerts: probe.alerts?.map((alert) => {
-          if (alert.query) {
-            showDeprecateMsg.query = true
-            return { ...alert, assertion: alert.query }
-          }
-
-          return alert
-        }),
-      })),
-    }
-
-    if (showDeprecateMsg.query) {
-      log.warn('"alerts.query" is deprecated. Please use "alerts.assertion"')
-    }
-
-    return checkedConfig
-  }
-
-  getProbes({ config, flags }: GetProbesParams): Probe[] {
-    const sortedProbes = sortProbes(config.probes, flags.id)
-
-    return sortedProbes.map((probe: Probe) =>
-      sanitizeProbe(isSymonModeFrom(flags), probe)
-    )
-  }
-
   async run(): Promise<void> {
-    const monika = await this.parse(Monika)
-    const _flags: MonikaFlags = monika.flags
-    setContext({ flags: _flags })
+    const cmd = await this.parse(Monika)
+    const flags = sanitizeFlags(cmd.flags)
+
+    setContext({ flags, userAgent: this.config.userAgent })
 
     try {
-      if (_flags['create-config']) {
-        await createConfig(_flags)
+      if (flags['create-config']) {
+        await createConfig(flags)
         return
       }
 
-      await openLogfile()
-
-      if (_flags.logs) {
+      if (flags.logs) {
         await printAllLogs()
-        await closeLog()
         return
       }
 
-      if (_flags.flush) {
-        await flush(_flags.force)
-        await closeLog()
+      if (flags.flush) {
+        await flush()
         return
       }
 
-      if (_flags.summary) {
-        printSummary(this.config)
-        await closeLog()
+      if (flags.summary) {
+        await printSummary()
         return
       }
 
-      await initLoaders(_flags, this.config)
+      await initLoaders(flags, this.config)
 
-      if (isSymonModeFrom(_flags)) {
-        symonClient = new SymonClient(_flags)
+      const isSymonMode = isSymonModeFrom(flags)
+      await logRunningInfo({ isSymonMode, isVerbose: flags.verbose })
+
+      if (isSymonMode) {
+        symonClient = new SymonClient(flags)
         await symonClient.initiate()
       }
 
       let isFirstRun = true
 
       for (;;) {
-        const controller = new AbortController()
-        const { signal } = controller
         const config = getConfig()
-        const notifications = config.notifications || []
-        const probes = this.getProbes({ config, flags: _flags })
+        const probes = getProbes({ config, flags })
 
         // emit the sanitized probe
         em.emit(events.config.sanitized, probes)
-
         // save some data into files for later
-        savePidFile(_flags.config, config)
-
-        this.deprecationHandler(config)
-
+        savePidFile(flags.config, config)
+        deprecationHandler(config)
         logStartupMessage({
           config,
-          flags: _flags,
+          flags,
           isFirstRun,
         })
 
+        const controller = new AbortController()
+        const { signal } = controller
+        const notifications = config.notifications || []
         startProbing({
           notifications,
           probes,
@@ -417,9 +155,8 @@ export default class Monika extends Command {
         sendMonikaStartMessage(notifications).catch((error) =>
           log.error(error.message)
         )
-
         // schedule status update notification
-        scheduleSummaryNotification({ config, flags: _flags })
+        scheduleSummaryNotification({ config, flags })
 
         isFirstRun = false
 
@@ -429,10 +166,82 @@ export default class Monika extends Command {
         controller.abort('Monika configuration updated')
       }
     } catch (error: unknown) {
-      await closeLog()
       this.error((error as Error)?.message, { exit: 1 })
+    } finally {
+      await closeLog()
     }
   }
+}
+
+type RunningInfoParams = { isSymonMode: boolean; isVerbose: boolean }
+
+async function logRunningInfo({ isVerbose, isSymonMode }: RunningInfoParams) {
+  if (!isVerbose && !isSymonMode) {
+    log.info('Monika is running.')
+    return
+  }
+
+  try {
+    const { city, hostname, isp, privateIp, publicIp } =
+      await fetchAndCacheNetworkInfo()
+
+    log.info(
+      `Monika is running from: ${city} - ${isp} (${publicIp}) - ${hostname} (${privateIp})`
+    )
+  } catch (error) {
+    log.warn(`Failed to obtain location/ISP info. Got: ${error}`)
+  }
+}
+
+type GetProbesParams = {
+  config: Config
+  flags: MonikaFlags
+}
+
+function getProbes({ config, flags }: GetProbesParams): Probe[] {
+  const sortedProbes = sortProbes(config.probes, flags.id)
+
+  return sortedProbes.map((probe: Probe) =>
+    sanitizeProbe(isSymonModeFrom(flags), probe)
+  )
+}
+
+function deprecationHandler(config: Config): Config {
+  const showDeprecateMsg: Record<'query', boolean> = {
+    query: false,
+  }
+
+  const checkedConfig = {
+    ...config,
+    probes: config.probes?.map((probe) => ({
+      ...probe,
+      requests: probe.requests?.map((request) => ({
+        ...request,
+        alert: request.alerts?.map((alert) => {
+          if (alert.query) {
+            showDeprecateMsg.query = true
+            return { ...alert, assertion: alert.query }
+          }
+
+          return alert
+        }),
+      })),
+      alerts: probe.alerts?.map((alert) => {
+        if (alert.query) {
+          showDeprecateMsg.query = true
+          return { ...alert, assertion: alert.query }
+        }
+
+        return alert
+      }),
+    })),
+  }
+
+  if (showDeprecateMsg.query) {
+    log.warn('"alerts.query" is deprecated. Please use "alerts.assertion"')
+  }
+
+  return checkedConfig
 }
 
 /**
